fix add with same argumengs (relational add)
add %eax %eax e.g. does not need to be top if eax is top.
It should be eax << 1. Sadly we sometimes have to guess the region.
This also improves precision for add foo foo in generals - addition is
fully relational in this case.
todo:
replace shift version that works only for 2^n operands to + by
mulSingleton-based one...